Fix date size for windows

+++ ktor-utils/mingwX64/src/io/ktor/util/date/DateMingw.kt
@@ -7,6 +7,6 @@ package io.ktor.util.date
 import kotlinx.cinterop.*
 import platform.posix.*
 
-internal actual fun system_time(tm: CValuesRef<tm>?): Int {
+internal actual fun system_time(tm: CValuesRef<tm>?): Long {
     return _mkgmtime(tm).convert()
 }
